+++ database/migrations/2023_12_08_135307_create_courses_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('name', 100);
             $table->text('description');
             $table->tinyInteger('level');
-            $table->float('price');
-            $table->float('promotional_price')->default(0);
+            $table->bigInteger('price');
+            $table->bigInteger('promotional_price')->default(0);
             $table->string('created_by', 100)->default('');
             $table->string('updated_by', 100)->default('');
             $table->unsignedBiginteger('subject_id')->unsigned();
